changed some details

+++ CatalogAPI/Models/Product.cs
@@ -14,7 +14,7 @@ namespace CatalogAPI.Models
         public double Price { get; set; }
         public int RAM { get; set; }
         public int Year { get; set; }
-        public int Display { get; set; }
+        public decimal Display { get; set; }
         public int Battery { get; set; }
         public int Camera { get; set; }
         public string Image { get; set; }
